drftutorial/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.5
django-oauth-toolkit==1.0.0
djangorestframework==3.6.4
httpie==0.9.9
idna==2.6
oauthlib==2.0.4
pkg-resources==0.0.0
Pygments==2.2.0
PyJWT==1.5.3
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
urllib3==1.22
